Adapt the reformed db hierarchy
