+++ origin/origin_ui.go
@@ -24,11 +24,11 @@ import (
 	"time"
 
 	"github.com/gin-gonic/gin"
-	"github.com/google/uuid"
 	"github.com/pkg/errors"
 	log "github.com/sirupsen/logrus"
 
 	"github.com/pelicanplatform/pelican/config"
+	"github.com/pelicanplatform/pelican/database"
 	"github.com/pelicanplatform/pelican/param"
 	"github.com/pelicanplatform/pelican/server_structs"
 	"github.com/pelicanplatform/pelican/server_utils"
@@ -53,15 +53,6 @@ type (
 		HttpServiceUrl string `json:"httpServiceUrl,omitempty"`
 
 		Exports []exportWithStatus `json:"exports"`
-	}
-
-	DowntimeInput struct {
-		CreatedBy   string                  `json:"createdBy"` // Person who created this downtime
-		Class       server_structs.Class    `json:"class"`
-		Description string                  `json:"description"`
-		Severity    server_structs.Severity `json:"severity"`
-		StartTime   int64                   `json:"startTime"` // Epoch UTC in seconds
-		EndTime     int64                   `json:"endTime"`   // Epoch UTC in seconds
 	}
 )
 
@@ -171,157 +162,6 @@ func handleExports(ctx *gin.Context) {
 	ctx.JSON(http.StatusOK, res)
 }
 
-func isValidClass(class server_structs.Class) bool {
-	validClasses := map[server_structs.Class]bool{
-		server_structs.SCHEDULED:   true,
-		server_structs.UNSCHEDULED: true,
-	}
-	return validClasses[class]
-}
-
-func isValidSeverity(severity server_structs.Severity) bool {
-	validSeverities := map[server_structs.Severity]bool{
-		server_structs.Outage:                      true,
-		server_structs.Severe:                      true,
-		server_structs.IntermittentOutage:          true,
-		server_structs.NoSignificantOutageExpected: true,
-	}
-	return validSeverities[severity]
-}
-
-func handleCreateDowntime(ctx *gin.Context) {
-	var downtimeInput DowntimeInput
-	if err := ctx.ShouldBindJSON(&downtimeInput); err != nil {
-		ctx.JSON(http.StatusBadRequest, server_structs.SimpleApiResp{Status: server_structs.RespFailed, Msg: "Invalid downtime request payload"})
-		return
-	}
-
-	id, err := uuid.NewV7()
-	if err != nil {
-		ctx.JSON(http.StatusInternalServerError, server_structs.SimpleApiResp{
-			Status: server_structs.RespFailed,
-			Msg:    "Failed to create new UUID for new entry in downtimes table",
-		})
-		return
-	}
-
-	if !isValidClass(downtimeInput.Class) {
-		ctx.JSON(http.StatusBadRequest, server_structs.SimpleApiResp{Status: server_structs.RespFailed, Msg: "Invalid downtime class"})
-		return
-	}
-	if !isValidSeverity(downtimeInput.Severity) {
-		ctx.JSON(http.StatusBadRequest, server_structs.SimpleApiResp{Status: server_structs.RespFailed, Msg: "Invalid downtime severity"})
-		return
-	}
-
-	downtime := server_structs.Downtime{
-		UUID:        id.String(),
-		CreatedBy:   downtimeInput.CreatedBy,
-		Class:       downtimeInput.Class,
-		Description: downtimeInput.Description,
-		Severity:    downtimeInput.Severity,
-		StartTime:   downtimeInput.StartTime,
-		EndTime:     downtimeInput.EndTime,
-	}
-
-	if err := createDowntime(&downtime); err != nil {
-		ctx.JSON(http.StatusInternalServerError, server_structs.SimpleApiResp{Status: server_structs.RespFailed, Msg: "Failed to create downtime: " + err.Error()})
-		return
-	}
-	ctx.JSON(http.StatusOK, downtime)
-}
-
-func handleGetActiveDowntime(ctx *gin.Context) {
-	downtime, err := getActiveDowntimes()
-	if err != nil {
-		ctx.JSON(http.StatusNotFound, server_structs.SimpleApiResp{Status: server_structs.RespFailed, Msg: "Failed to get active downtime: " + err.Error()})
-		return
-	}
-	ctx.JSON(http.StatusOK, downtime)
-}
-
-func handleGetAllDowntime(ctx *gin.Context) {
-	downtime, err := getAllDowntimes()
-	if err != nil {
-		ctx.JSON(http.StatusNotFound, server_structs.SimpleApiResp{Status: server_structs.RespFailed, Msg: "Failed to get all downtime: " + err.Error()})
-		return
-	}
-	ctx.JSON(http.StatusOK, downtime)
-}
-
-func handleGetDowntimeByUUID(ctx *gin.Context) {
-	uuid := ctx.Param("uuid")
-	downtime, err := getDowntimeByUUID(uuid)
-	if err != nil {
-		ctx.JSON(http.StatusNotFound, server_structs.SimpleApiResp{Status: server_structs.RespFailed, Msg: "Failed to get downtime by UUID: " + err.Error()})
-		return
-	}
-	ctx.JSON(http.StatusOK, downtime)
-}
-
-func handleUpdateDowntime(ctx *gin.Context) {
-	uuid := ctx.Param("uuid")
-	var downtimeInput DowntimeInput
-	if err := ctx.ShouldBindJSON(&downtimeInput); err != nil {
-		ctx.JSON(http.StatusBadRequest, server_structs.SimpleApiResp{Status: server_structs.RespFailed, Msg: "Invalid request payload"})
-		return
-	}
-
-	if downtimeInput.Class != "" && !isValidClass(downtimeInput.Class) {
-		ctx.JSON(http.StatusBadRequest, server_structs.SimpleApiResp{Status: server_structs.RespFailed, Msg: "Invalid input downtime class"})
-		return
-	}
-	if downtimeInput.Severity != "" && !isValidSeverity(downtimeInput.Severity) {
-		ctx.JSON(http.StatusBadRequest, server_structs.SimpleApiResp{Status: server_structs.RespFailed, Msg: "Invalid input downtime severity"})
-		return
-	}
-
-	// Retrieve existing downtime record from the database
-	existingDowntime, err := getDowntimeByUUID(uuid)
-	if err != nil {
-		ctx.JSON(http.StatusNotFound, server_structs.SimpleApiResp{
-			Status: server_structs.RespFailed,
-			Msg:    "Downtime record not found",
-		})
-		return
-	}
-
-	// Only update fields provided in the request
-	if downtimeInput.CreatedBy != "" {
-		existingDowntime.CreatedBy = downtimeInput.CreatedBy
-	}
-	if downtimeInput.Class != "" {
-		existingDowntime.Class = downtimeInput.Class
-	}
-	if downtimeInput.Description != "" {
-		existingDowntime.Description = downtimeInput.Description
-	}
-	if downtimeInput.Severity != "" {
-		existingDowntime.Severity = downtimeInput.Severity
-	}
-	if downtimeInput.StartTime != 0 {
-		existingDowntime.StartTime = downtimeInput.StartTime
-	}
-	if downtimeInput.EndTime != 0 {
-		existingDowntime.EndTime = downtimeInput.EndTime
-	}
-
-	if err := updateDowntime(uuid, existingDowntime); err != nil {
-		ctx.JSON(http.StatusInternalServerError, server_structs.SimpleApiResp{Status: server_structs.RespFailed, Msg: "Failed to update downtime: " + err.Error()})
-		return
-	}
-	ctx.JSON(http.StatusOK, existingDowntime)
-}
-
-func handleDeleteDowntime(ctx *gin.Context) {
-	uuid := ctx.Param("uuid")
-	if err := deleteDowntime(uuid); err != nil {
-		ctx.JSON(http.StatusInternalServerError, server_structs.SimpleApiResp{Status: server_structs.RespFailed, Msg: "Failed to delete downtime: " + err.Error()})
-		return
-	}
-	ctx.JSON(http.StatusOK, server_structs.SimpleApiResp{Status: server_structs.RespOK, Msg: "Downtime deleted successfully"})
-}
-
 func RegisterOriginWebAPI(engine *gin.Engine) error {
 	originWebAPI := engine.Group("/api/v1.0/origin_ui", web_ui.ServerHeaderMiddleware)
 	{
@@ -330,12 +170,12 @@ func RegisterOriginWebAPI(engine *gin.Engine) error {
 
 	originDowntimeAPI := originWebAPI.Group("/downtime")
 	{
-		originDowntimeAPI.POST("/", web_ui.AuthHandler, web_ui.AdminAuthHandler, handleCreateDowntime)
-		originDowntimeAPI.GET("/", web_ui.AuthHandler, web_ui.AdminAuthHandler, handleGetActiveDowntime)
-		originDowntimeAPI.GET("/all", web_ui.AuthHandler, web_ui.AdminAuthHandler, handleGetAllDowntime)
-		originDowntimeAPI.GET("/:uuid", web_ui.AuthHandler, web_ui.AdminAuthHandler, handleGetDowntimeByUUID)
-		originDowntimeAPI.PUT("/:uuid", web_ui.AuthHandler, web_ui.AdminAuthHandler, handleUpdateDowntime)
-		originDowntimeAPI.DELETE("/:uuid", web_ui.AuthHandler, web_ui.AdminAuthHandler, handleDeleteDowntime)
+		originDowntimeAPI.POST("/", web_ui.AuthHandler, web_ui.AdminAuthHandler, database.HandleCreateDowntime)
+		originDowntimeAPI.GET("/", web_ui.AuthHandler, web_ui.AdminAuthHandler, database.HandleGetActiveDowntime)
+		originDowntimeAPI.GET("/all", web_ui.AuthHandler, web_ui.AdminAuthHandler, database.HandleGetAllDowntime)
+		originDowntimeAPI.GET("/:uuid", web_ui.AuthHandler, web_ui.AdminAuthHandler, database.HandleGetDowntimeByUUID)
+		originDowntimeAPI.PUT("/:uuid", web_ui.AuthHandler, web_ui.AdminAuthHandler, database.HandleUpdateDowntime)
+		originDowntimeAPI.DELETE("/:uuid", web_ui.AuthHandler, web_ui.AdminAuthHandler, database.HandleDeleteDowntime)
 	}
 
 	// Globus backend specific. Config other origin routes above this line
